Revisions to update handling (1 of ???)
Pushing changes to move the update handling back to the library layout,
this will simplify the plugin even though it'll be a single plugin
accessing said library.  (This is for clarity, don't really want the
update library being used by more than this plugin)

Anyways, interrim commit to preserve work to date.